[feature] first version of test ui application

+++ libraries/core/include/core/filesystem.h
@@ -13,10 +13,10 @@ namespace rpp
      */
     typedef u32 FileHandle;
 
-#define FILE_MODE_READ 0x00       ///< Open the file for reading.
-#define FILE_MODE_WRITE 0x01      ///< Open the file for writing (overwrites existing content).
-#define FILE_MODE_APPEND 0x02     ///< Open the file for appending (adds to the end of the file).
-#define FILE_MODE_READ_WRITE 0x03 ///< Open the file for both reading and writing.
+#define FILE_MODE_READ u32(0x00)       ///< Open the file for reading.
+#define FILE_MODE_WRITE u32(0x01)      ///< Open the file for writing (overwrites existing content).
+#define FILE_MODE_APPEND u32(0x02)     ///< Open the file for appending (adds to the end of the file).
+#define FILE_MODE_READ_WRITE u32(0x03) ///< Open the file for both reading and writing.
 
     /**
      * The file system module provides functionalities for file and directory operations.
@@ -94,7 +94,7 @@ namespace rpp
          * the function checks if the file can be opened for writing.
          * @return The id of the file if it exists, or INVALID_ID if it does not.
          */
-        static FileHandle OpenFile(const String &filePath, u8 mode = FILE_MODE_READ) RPP_E2E_BINDING;
+        static FileHandle OpenFile(const String &filePath, u32 mode = FILE_MODE_READ) RPP_E2E_BINDING;
 
         /**
          * @brief Check if current file id is opened or not.
